+++ app/src/main/java/com/sopt/now/data/Friend/FriendDao.kt
@@ -1,9 +1,10 @@
-package com.sopt.now.data
+package com.sopt.now.data.Friend
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import com.sopt.now.data.Friend.Friend
 import kotlinx.coroutines.flow.Flow
 
 @Dao
