+++ app/src/main/java/com/sopt/now/ui/home/FriendViewHolder.kt
@@ -1,7 +1,7 @@
 package com.sopt.now.ui.home
 
 import androidx.recyclerview.widget.RecyclerView
-import com.sopt.now.data.Friend
+import com.sopt.now.data.Friend.Friend
 import com.sopt.now.databinding.ItemFriendBinding
 
 class FriendViewHolder(private val binding: ItemFriendBinding) :
