+++ app/src/main/java/com/sopt/now/ui/home/HomeViewModel.kt
@@ -8,9 +8,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import com.sopt.now.R
-import com.sopt.now.data.AppContainer
-import com.sopt.now.data.Friend
-import com.sopt.now.data.FriendsRepository
+import com.sopt.now.data.Friend.AppContainer
+import com.sopt.now.data.Friend.Friend
+import com.sopt.now.data.Friend.FriendsRepository
 import com.sopt.now.data.Profile
 import kotlinx.coroutines.launch
